VideoXplorer/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.14
azure-nspkg==2.0.0
azure-storage-blob==1.3.1
azure-storage-common==1.3.0
azure-storage-nspkg==3.0.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
idna==2.7
numpy==1.15.0
pycparser==2.18
pydocumentdb==2.3.2
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
opencv-python==3.4.2.17